@@ -1,6 +1,6 @@
 {
 	"name": "skynovel2",
-	"version": "4.25.3",
+	"version": "4.26.0",
 	"publisher": "famibee2",
 	"displayName": "SKYNovel",
 	"description": "It supports development of SKYNovel for VSCode, VSCodeでSKYNovelの開発をサポートする拡張機能",
@@ -9,7 +9,7 @@
 	"homepage": "http://famibee.blog38.fc2.com/",
 	"license": "MIT",
 	"engines": {
-		"vscode": "^1.104.0"
+		"vscode": "^1.105.0"
 	},
 	"categories": [
 		"Programming Languages",
@@ -695,16 +695,17 @@
 		"@vscode/debugprotocol": "^1.68.0",
 		"adm-zip": "^0.5.16",
 		"archiver": "^7.0.1",
-		"await-notify": "^1.0.1",
 		"crc-32": "^1.2.2",
 		"encoding-japanese": "^2.2.0",
+		"fluent-ffmpeg": "^2.1.3",
 		"fs-extra": "^11.3.2",
 		"image-size": "^2.0.2",
-		"npm-check-updates": "^19.0.0",
+		"npm-check-updates": "^19.1.0",
 		"original-fs": "^1.2.0",
 		"pinia": "^3.0.3",
+		"psd.js": "^3.9.2",
 		"socket.io": "^4.8.1",
-		"string-replace-async": "^3.0.2",
+		"str-async-replace": "^1.1.2",
 		"uuid": "^13.0.0",
 		"vee-validate": "^4.15.1",
 		"vscode-languageclient": "^9.0.1",
@@ -712,55 +713,50 @@
 		"yup": "^1.7.1"
 	},
 	"devDependencies": {
-		"@ffmpeg-installer/ffmpeg": "^1.1.0",
+		"@eslint/js": "^9.37.0",
 		"@types/adm-zip": "^0.5.7",
 		"@types/archiver": "^6.0.3",
 		"@types/bootstrap": "^5.2.10",
 		"@types/encoding-japanese": "^2.2.1",
+		"@types/eslint-scope": "^8.3.2",
 		"@types/fluent-ffmpeg": "^2.1.27",
 		"@types/fs-extra": "^11.0.4",
 		"@types/jest": "^30.0.0",
 		"@types/minimatch": "^6.0.0",
-		"@types/node": "^24.5.2",
+		"@types/node": "^24.7.2",
 		"@types/power-assert": "^1.5.12",
 		"@types/psd": "^3.4.3",
 		"@types/sharp": "^0.32.0",
-		"@types/vscode": "^1.104.0",
+		"@types/vscode": "^1.105.0",
 		"@types/vscode-webview": "^1.57.5",
-		"@typescript-eslint/eslint-plugin": "^8.44.1",
-		"@typescript-eslint/parser": "^8.44.1",
 		"@vitejs/plugin-vue": "^6.0.1",
 		"@vscode/test-electron": "^2.5.2",
-		"electron": "^38.2.0",
-		"esbuild": "^0.25.10",
-		"file-loader": "^6.2.0",
-		"p-queue": "^8.1.1",
+		"electron": "^38.3.0",
+		"esbuild": "^0.25.11",
+		"eslint": "^9.37.0",
+		"eslint-plugin-import": "^2.32.0",
+		"globals": "^16.4.0",
 		"png2icons": "^2.0.1",
 		"rimraf": "^6.0.1",
-		"terser": "^5.44.0",
 		"ts-loader": "^9.5.4",
-		"typescript": "^5.9.2",
-		"vite": "^7.1.7",
+		"typescript": "^5.9.3",
+		"typescript-eslint": "^8.46.1",
+		"vite": "^7.1.10",
 		"vscode-languageserver-types": "^3.17.5",
-		"vue-loader": "^17.4.2",
-		"webpack": "^5.101.3",
+		"webpack": "^5.102.1",
 		"webpack-cli": "^6.0.1"
-	},
-	"optionalDependencies": {
-		"bufferutil": "^4.0.9",
-		"utf-8-validate": "^6.0.5"
 	},
 	"scripts": {
 		"build": "bun vue_bld && bun build.ts",
-		"watch": "bun --watch build.ts --watch",
+		"watch": "bun vue_bw & bun --watch build.ts --watch",
 		"md2json": "bun src/md2json.ts && bun build:types",
 		"build:types": "tsc -P ./src/tsconfig.types.json",
-		"vue_prv": "vite -c ./views/vite.config.mts",
 		"vue_bld": "vite build -c ./views/vite.config.mts",
+		"vue_bw": "vite build -w -c ./views/vite.config.mts",
 		"update": "bun update && bun run build && bun ncu_lsp && ncu -g",
 		"ncu_lsp": "cd server && bun update && bun outdated && cd ..",
 		"postinstall": "cd server && bun i && cd ..",
-		"rebuild": "rimraf node_modules && bun pm cache rm && bun i",
+		"rebuild": "rimraf package-lock.json && rimraf node_modules && bun pm cache rm && bun i",
 		"version:patch": "npm -no-git-tag-version version patch",
 		"version:minor": "npm -no-git-tag-version version minor",
 		"version:major": "npm -no-git-tag-version version major",
