@@ -7,8 +7,8 @@
 
 
 // =============== Global
-export function int(o: any): number {return parseInt(String(o), 10)}
-export function uint(o: any): number {
+export function int(o: unknown): number {return parseInt(String(o), 10)}
+export function uint(o: unknown): number {
 	const v = parseInt(String(o), 10);
 	return v < 0 ? -v : v;
 }
@@ -18,6 +18,15 @@ export	const	REG_SCRIPT	= /\.ss?n$/;
 export	const docsel = {scheme: 'file', language: 'skynovel'};
 
 
+// =============== WebView
+import type {Uri, Webview} from 'vscode';
+export function repWvUri(inp: string, wv: Webview, uriDoc: Uri): string {
+	return inp
+	.replaceAll('${webview.cspSource}', wv.cspSource)
+	.replaceAll(/(href|src)="\.\//g, `$1="${String(wv.asWebviewUri(uriDoc))}/`);	// 必ず String() で
+}
+
+
 // =============== Project
 export const is_win = process.platform === 'win32';
 export const is_mac = process.platform === 'darwin';
@@ -25,10 +34,36 @@ export const is_mac = process.platform === 'darwin';
 
 
 // =============== LSP
-import {FULL_PATH, FULL_SCH_PATH} from '../server/src/LspWs';
-export function fsp2fp(s: FULL_SCH_PATH): FULL_PATH {return s.replace(/(?:\/\w:)?/, '');}
+export const REQ_ID = ':SKYNovel:';	// これは server/src/LangSrv.ts に置くの禁止
+
+
+export type WORKSPACE_PATH	= string;	// doc/prj/script/main.sn
+export type PROJECT_PATH	= string;	// script/main.sn
+export type FULL_PATH		= string;	// /[user]/.../[prj]/doc/prj/script/main.sn
+export type VSC_FULL_PATH	= string;	// c:\[user]\...\[prj]\doc\prj\script\main.sn
+export type FULL_SCH_PATH	= string;	// file://c:\[user]\...\[prj]/doc/prj/
+										// scheme つき
+
+export function vsc2fp(p: VSC_FULL_PATH): FULL_PATH {return p.replace(/(?:\/\w:)?/, '');}
 	// FULL_SCH_PATH は uri.path など
 	// 4win 先頭の【'/'+ ドライブ名（小文字）】を取って扱う用
+
+export function fullSchPath2fp(fsp: FULL_SCH_PATH): FULL_PATH {
+	return decodeURIComponent(fsp.replace(/file:\/\/(\/\w%3A)?/, ''));
+}	// 似たような名前のメソッドになるので目立たせる
+	// 逆方向は難しそう、変換前の値は保存必要か
+
+//	docs.get(fsp) などにはこれが必要
+//NOTE: 雑コード
+export const fp2fullSchPath: (fp: FULL_PATH)=> FULL_SCH_PATH = is_win
+	? fp=> 'file://c:'+ encodeURI(fp)
+	: fp=> 'file://'+ encodeURI(fp);
+
+
+export function uri2path(p: string): string {return p.slice(7)}
+	// 'file://' を取る
+
+
 /*
 // console.log(`fn:Project.ts drop scheme:${scheme} fp:${fp}: uri:${uri.toString()}: path=${path}= fsPath-${uri.fsPath}-`);
 
@@ -48,13 +83,8 @@ fn:Project.ts drop scheme:file
 		> ファイルI/Oにおいて、パス文字列の先頭に「\\?\」を付けると、Windows APIは文字列解析を一切行わず、それに続く文字列をファイルシステムに直接送信します。例えば、ファイルシステムが長いパスとファイル名をサポートしている場合、 Windows APIによって強制されるMAX_PATHの制限を超えることができます。
 */
 
-export type T_DIAG_L2S = {
-	mes: string,
-	sev: 'E'|'W'|'I'|'H',
-};
-export type T_H_ADIAG_L2S = {[fp: string]: T_DIAG_L2S[]};
-
-export const hDiagL2s	:{[code_name: string]: T_DIAG_L2S} = {
+import type {T_H_ADIAG} from '../server/src/LspWs';
+export const hDiagL2s	:{[code_name: string]: T_H_ADIAG} = {
 	ファイル重複: {
 		mes	: 'プロジェクト内でファイル【$】が重複しています。フォルダを縦断検索するため許されません',
 		sev	: 'W',
@@ -64,7 +94,7 @@ export const hDiagL2s	:{[code_name: string]: T_DIAG_L2S} = {
 		sev	: 'W',
 	},
 	文字コード異常: {
-		mes	: `文字コードが異常（$）です。UTF8 か ASCII にして下さい`,
+		mes	: '文字コードが異常（$）です。UTF8 か ASCII にして下さい',
 		sev	: 'E',
 	},
 }
@@ -74,12 +104,12 @@ export const hDiagL2s	:{[code_name: string]: T_DIAG_L2S} = {
 import {basename, extname, resolve} from 'node:path';
 import {readdirSync, existsSync, readFileSync, ensureFileSync, statSync, writeFileSync} from 'fs-extra';
 
-const REG_SYS_FN = /^(_notes|Icon\r|\.[^\/]+|[^\/]+\.(db|ini|git))$/;
+const REG_SYS_FN = /^(_notes|Icon\r|\.[^/]+|[^/]+\.(db|ini|git))$/;
 	// 6 matches (144 steps, 0.1ms)【\n 入注意】 https://regex101.com/r/uFkUrb/1
 
 export function treeProc(wd: FULL_SCH_PATH, fnc: (fp: FULL_PATH)=> void) {
 	for (const d of readdirSync(wd, {withFileTypes: true})) {
-		const nm = String(d.name).normalize('NFC');
+		const nm = d.name.normalize('NFC');
 		if (REG_SYS_FN.test(nm)) continue;
 		const fp = resolve(wd, nm);
 		if (d.isDirectory()) {treeProc(fp, fnc); continue;}
@@ -90,7 +120,7 @@ export function treeProc(wd: FULL_SCH_PATH, fnc: (fp: FULL_PATH)=> void) {
 
 export function foldProc(wd: FULL_SCH_PATH, fnc: (fp: FULL_PATH, nm: string)=> void, fncFld: (nm: string)=> void) {
 	for (const d of readdirSync(wd, {withFileTypes: true})) {
-		const nm = String(d.name).normalize('NFC');
+		const nm = d.name.normalize('NFC');
 		if (REG_SYS_FN.test(nm)) continue;
 		if (d.isDirectory()) {fncFld(nm); continue;}
 
@@ -108,13 +138,13 @@ export function replaceFile(src: FULL_SCH_PATH, r: RegExp, rep: string, verbose 
 		if (dest !== src) ensureFileSync(dest);	// これが無いとエラーになったので
 
 		const txt = readFileSync(src, {encoding: 'utf8'});
-		const ret = String(txt.replace(r, rep));
+		const ret = txt.replace(r, rep);
 		if (txt !== ret) {writeFileSync(dest, ret); return true}
 
 		if (verbose) console.error(`replaceFile fail by same:${src}`);
 	}
-	catch (err) {
-		console.error(`replaceFile src:${src} ${err}`);
+	catch (e) {
+		console.error(`replaceFile src:${src} %o`, e);
 	}
 	return false;
 }
@@ -134,8 +164,8 @@ export function replaceRegsFile(src: FULL_SCH_PATH, a: [r: RegExp, rep: string][
 
 		if (verbose) console.error(`replaceRegsFile fail by same:${src}`);
 	}
-	catch (err) {
-		console.error(`replaceRegsFile src:${src} ${err}`);
+	catch (e) {
+		console.error(`replaceRegsFile src:${src} %o`, e);
 	}
 	return false;
 }
@@ -158,14 +188,15 @@ export function chkUpdate(path1: FULL_PATH, path2: FULL_PATH, doesnt_exist = tru
 
 	return hash[name] = chkBoolean(hash[name]);
 }*/
-export	function chkBoolean(v: any): boolean {
+export	function chkBoolean(v: unknown): boolean {
 	if (v === null) return false;
 
+	// eslint-disable-next-line @typescript-eslint/no-base-to-string
 	const v2 = String(v);
 	return (v2 === 'false')? false : Boolean(v2);
 }
 
-export	function getFn(path: string) {return basename(path, extname(path))};
+export	function getFn(path: string) {return basename(path, extname(path))}
 
 
 // =============== EncryptorTransform
@@ -173,12 +204,11 @@ export type HArg = {
 	タグ名?	: string;
 
 	layer?	: string;	// レイヤ系
+	class?	: string;
 };
-export interface ITag { (hArg: HArg): boolean; }
+export type ITag = (hArg: HArg)=> boolean;
 
-export interface ILayerFactory {
-	(): any;
-}
+export type ILayerFactory = ()=> unknown;
 
 export type PLUGIN_DECAB_RET = {
 	ext_num	: number;
