@@ -22,7 +22,7 @@ export const enum SEARCH_PATH_ARG_EXT {	// #searchPath 使用時、第二引数�
 	TST_EEE		= 'eee',
 	TST_GGG		= 'ggg',
 	TST_PNGXML	= 'png|xml',
-};
+}
 
 
 export type T_CFG = {
@@ -100,19 +100,24 @@ export	const DEF_CFG: T_CFG = {
 };
 
 
-
-export interface IExts { [ext: string]: string; };
-export interface IFn2Path { [fn: string]: IExts; };
+export type IExts = {
+	':cnt'?			: number;
+} & {
+	[ext: string]	: string;	// pp
+};
+// export type IExts = { [ext: string]: string; };
+export type IFn2Path = { [fn: string]: IExts; };
 
 export type T_SEARCHPATH = (fn: string, extptn?: SEARCH_PATH_ARG_EXT)=> string;
-export interface IConfig {
+export type IConfig = {
 	oCfg	: T_CFG;
 	getNs()	: string;
 	searchPath: T_SEARCHPATH;
-	matchPath(fnptn: string, extptn?: SEARCH_PATH_ARG_EXT): ReadonlyArray<IExts>;
+	matchPath(fnptn: string, extptn?: SEARCH_PATH_ARG_EXT): readonly IExts[];
 	addPath(fn: string, h_exts: IExts): void;
 }
 
+// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
 export interface ISysRoots {
 	loadPath(hPathFn2Exts: IFn2Path, cfg: IConfig): Promise<void>;
 	dec(ext: string, tx: string): Promise<string>;
@@ -130,7 +135,7 @@ export type HSysBaseArg = {
 
 
 export class ConfigBase implements IConfig {
-	oCfg	= DEF_CFG;
+	oCfg	= structuredClone(DEF_CFG);
 
 	userFnTail		= '';	// 4tst public
 	protected	hPathFn2Exts	: IFn2Path	= {};
@@ -138,14 +143,14 @@ export class ConfigBase implements IConfig {
 	protected	constructor(readonly sys: ISysRoots) {}
 	async load(oCfg: T_CFG) {
 		// this.oCfg = {...this.oCfg, ...oCfg};	// 一階層目でコピーしてしまう
-		this.oCfg.save_ns = oCfg?.save_ns ?? this.oCfg.save_ns;
+		this.oCfg.save_ns = oCfg.save_ns || this.oCfg.save_ns;
 
-		this.oCfg.window.width = Number(oCfg?.window?.width ?? this.oCfg.window.width);
-		this.oCfg.window.height = Number(oCfg?.window?.height ?? this.oCfg.window.height);
+		this.oCfg.window.width = oCfg.window.width || this.oCfg.window.width;
+		this.oCfg.window.height = oCfg.window.height || this.oCfg.window.height;
 
 		this.oCfg.book = {...this.oCfg.book, ...oCfg.book};
 
-		this.oCfg.log.max_len = oCfg.log?.max_len ?? this.oCfg.log.max_len;
+		this.oCfg.log.max_len = oCfg.log.max_len || this.oCfg.log.max_len;
 
 		this.oCfg.init = {...this.oCfg.init, ...oCfg.init};
 
@@ -172,13 +177,13 @@ export class ConfigBase implements IConfig {
 		}
 		else
 		for (const [fn0, hExts] of Object.entries(this.hPathFn2Exts)) {
-			for (const [ext, v] of Object.entries(hExts)) {
+			for (const [ext, pp] of Object.entries(hExts)) {
 				if (! ext.startsWith(':')) {
 					hFn2Ext[fn0] = ext;
 					continue;
 				}
 				if (! ext.endsWith(':id')) continue;
-				const hp = v.slice(v.lastIndexOf('/') +1);
+				const hp = (<string>pp).slice((<string>pp).lastIndexOf('/') +1);
 				const fn = hExts[ext.slice(0, -10)] ?? '';
 				const res = await this.sys.fetch(fn);
 				const src = await res.text();
@@ -194,19 +199,19 @@ export class ConfigBase implements IConfig {
 
 	getNs() {return `skynovel.${this.oCfg.save_ns} - `}
 
-	readonly	#REG_PATH = /([^\/\s]+)\.([^\d]\w+)/;
+	readonly	#REG_PATH = /([^/\s]+)\.([^\d]\w+)/;
 		// 4 match 498 step(~1ms)  https://regex101.com/r/tpVgmI/1
 	searchPath(fn: string, extptn: SEARCH_PATH_ARG_EXT = SEARCH_PATH_ARG_EXT.DEFAULT): string {
 		if (! fn) throw '[searchPath] fnが空です';
 		if (fn.startsWith('http://')) return fn;
 
 		const a = fn.match(this.#REG_PATH);
-		let fn0 = a ?a[1] :fn;
+		let fn0 = a ?a[1] ?? '' :fn;
 		const ext = a ?a[2] :'';
 		if (this.userFnTail) {
 			const utn = fn0 +'@@'+ this.userFnTail;
 			if (utn in this.hPathFn2Exts) {
-				if (extptn === '') fn0 = utn;
+				if (extptn === SEARCH_PATH_ARG_EXT.DEFAULT) fn0 = utn;
 				else for (const e3 of Object.keys(this.hPathFn2Exts[utn] ?? {})) {
 					if (! `|${extptn}|`.includes(`|${e3}|`)) continue;
 
@@ -215,13 +220,13 @@ export class ConfigBase implements IConfig {
 				}
 			}
 		}
-		const h_exts = this.hPathFn2Exts[fn0!];
+		const h_exts = this.hPathFn2Exts[fn0];
 		if (! h_exts) throw `サーチパスに存在しないファイル【${fn}】です`;
 
 		if (! ext) {	// fnに拡張子が含まれていない
 			//	extのどれかでサーチ（ファイル名サーチ→拡張子群にextが含まれるか）
 			const hcnt = int(h_exts[':cnt']);
-			if (extptn === '') {
+			if (extptn === SEARCH_PATH_ARG_EXT.DEFAULT) {
 				if (hcnt > 1) throw `指定ファイル【${fn}】が複数マッチします。サーチ対象拡張子群【${extptn}】で絞り込むか、ファイル名を個別にして下さい。`;
 
 				return fn;
@@ -235,15 +240,15 @@ export class ConfigBase implements IConfig {
 					if (++cnt > 1) throw `指定ファイル【${fn}】が複数マッチします。サーチ対象拡張子群【${extptn}】で絞り込むか、ファイル名を個別にして下さい。`;
 				}
 			}
-			for (const e of Object.keys(h_exts)) {
-				if (search_exts.includes(`|${e}|`)) return h_exts[e]!;
+			for (const [ext, pp] of Object.entries(h_exts)) {
+				if (search_exts.includes(`|${ext}|`)) return <string>pp;
 			}
 			throw `サーチ対象拡張子群【${extptn}】にマッチするファイルがサーチパスに存在しません。探索ファイル名=【${fn}】`;
 		}
 
 		// fnに拡張子xが含まれている
 		//	ファイル名サーチ→拡張子群にxが含まれるか
-		if (extptn !== '' && ! `|${extptn}|`.includes(`|${ext}|`)) {
+		if (extptn !== SEARCH_PATH_ARG_EXT.DEFAULT && ! `|${extptn}|`.includes(`|${ext}|`)) {
 			throw `指定ファイルの拡張子【${ext}】は、サーチ対象拡張子群【${extptn}】にマッチしません。探索ファイル名=【${fn}】`;
 		}
 
@@ -253,13 +258,13 @@ export class ConfigBase implements IConfig {
 		return ret;
 	}
 
-	matchPath(fnptn: string, extptn: SEARCH_PATH_ARG_EXT = SEARCH_PATH_ARG_EXT.DEFAULT): ReadonlyArray<IExts> {
+	matchPath(fnptn: string, extptn: SEARCH_PATH_ARG_EXT = SEARCH_PATH_ARG_EXT.DEFAULT): readonly IExts[] {
 		const aRet :IExts[] = [];
 		const regPtn = new RegExp(fnptn);
 		const regExt = new RegExp(extptn);
 		for (const [fn, h_exts] of Object.entries(this.hPathFn2Exts)) {
 			if (fn.search(regPtn) === -1) continue;
-			if (extptn === '') {aRet.push(h_exts); continue}
+			if (extptn === SEARCH_PATH_ARG_EXT.DEFAULT) {aRet.push(h_exts); continue}
 
 			const o :IExts = {};
 			let isa = false;
@@ -275,9 +280,9 @@ export class ConfigBase implements IConfig {
 	}
 
 	addPath(fn: string, h_exts: IExts) {
-		const o: any = {};
+		const o: IExts = {};
 		for (const [ext, v] of Object.entries(h_exts)) {
-			o[ext] = (ext.startsWith(':') ?`` :this.sys.arg.cur) + v;
+			o[ext] = (ext.startsWith(':') ?'' :this.sys.arg.cur) + String(v);
 		}
 		this.hPathFn2Exts[fn] = o;
 	}
