@@ -5,20 +5,21 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import type {T_E2V_INIT, T_E2V_CFG, T_E2V_NOTICE_COMPONENT, T_V2E_WSS, T_E2V} from '../views/types';
-import {DEF_WSS} from '../views/types';
-import {chkBoolean, replaceRegsFile, fsp2fp} from './CmnLib';
+import type {T_V2E_CFG, T_E2V_CFG, T_E2V, T_V2E, T_E2V_INIT, T_E2V_NOTICE_COMPONENT, TK_WSS} from './types';
+import {DEF_WSS} from './types';
+import {chkBoolean, replaceRegsFile, repWvUri} from './CmnLib';
 import {ActivityBar, eTreeEnv, getNonce} from './ActivityBar';
 import type {Config} from './Config';
 import {openURL} from './WorkSpaces';
-import type {WfbOptPic} from './WfbOptPic';
-import type {WfbOptSnd} from './WfbOptSnd';
-import type {WfbOptFont} from './WfbOptFont';
+import type {T_reqPrj2LSP} from './Project';
+import type {PrjCmn} from './PrjCmn';
+import type {WfbOptPic} from './batch/WfbOptPic';
+import type {WfbOptSnd} from './batch/WfbOptSnd';
+import type {WfbOptFont} from './batch/WfbOptFont';
 import {WPFolder} from './WPFolder';
-import {WfbSettingSn} from './WfbSettingSn';
-import type {HDiff} from './HDiff';
+import {WfbSettingSn} from './batch/WfbSettingSn';
 
-import type {Disposable, ExtensionContext, WebviewPanel, WorkspaceFolder} from 'vscode';
+import type {Disposable, WebviewPanel, WorkspaceFolder} from 'vscode';
 import {env, Uri, ViewColumn, window} from 'vscode';
 import {copyFile, ensureFile, existsSync, readFile, readFileSync, readJson, readJsonSync, writeFile} from 'fs-extra';
 import {basename} from 'path';
@@ -28,11 +29,9 @@ import {userInfo} from 'os';
 
 export class PrjSetting implements Disposable {
 	readonly	#wss;
-				#oWss			= DEF_WSS;
+				#oWss	= DEF_WSS;
 	get			oWss() {return this.#oWss}
 
-	readonly	#PATH_WS		: string;
-	readonly	#PATH_PRJ		: string;
 	readonly	#PATH_PRJ_JSON	: string;
 	readonly	#PATH_APP_JS	: string;
 	readonly	#PATH_PKG_JSON	: string;
@@ -40,7 +39,7 @@ export class PrjSetting implements Disposable {
 	readonly	#PATH_INS_NSH		: string;
 	readonly	#PATH_ICON			: string;
 	readonly	#PATH_README4FREEM	: string;
-	readonly	#localExtensionResRoots: Uri;
+	readonly	#uriRes: Uri;
 
 				#htmSrc	= '';
 
@@ -53,61 +52,51 @@ export class PrjSetting implements Disposable {
 //	readonly	#ds		= new DisposableStack;
 	readonly	#ds		: Disposable[]	= [];
 
-	readonly	#setEscape: ()=> void;
+	readonly	#setEscape: ()=> Promise<void>;
+
 
 	//MARK: コンストラクタ
 	constructor(
-		private readonly ctx: ExtensionContext,
-		readonly wsFld: WorkspaceFolder,
+		private readonly pc		: PrjCmn,
 		private readonly cfg	: Config,
-		private readonly setTitle: (title: string)=> void,
-				readonly sendRequest2LSP: (cmd: string, o?: any)=> void,
-				readonly diff	: HDiff,
+		private readonly setTitle	: (title: string)=> void,
+				readonly reqPrj2LSP	: T_reqPrj2LSP,
 		private readonly optPic	: WfbOptPic,
 		private readonly optSnd	: WfbOptSnd,
 		private readonly optFont: WfbOptFont,
-		private readonly FLD_SRC: string,
-		private readonly is_new_tmp	: boolean,
 	) {
-		this.#wss = ctx.workspaceState;
-		const oWss = DEF_WSS as {[nm: string]: any};
+		this.#wss = this.pc.ctx.workspaceState;
+		const oWss = DEF_WSS;
 		for (const [nm, v] of Object.entries(oWss)) {
 			const d = this.#wss.get(nm);
-			if (d) oWss[nm] = <any>d;
-			else /* await */ this.#wss.update(nm, v);
+			if (d) oWss[<TK_WSS>nm] = <never>d;
+			else void this.#wss.update(nm, v);
 		}
-		this.#oWss = oWss as any;
+		this.#oWss = oWss;
 
-		this.#PATH_WS = fsp2fp(wsFld.uri.path);
-		this.#PATH_PRJ = this.#PATH_WS +'/doc/prj/';
-		this.#PATH_PRJ_JSON = this.#PATH_PRJ +'prj.json';
-		this.#PATH_APP_JS = this.#PATH_WS +'/doc/app.js';
-		this.#PATH_PKG_JSON = this.#PATH_WS +'/package.json';
+		this.#PATH_PRJ_JSON = this.pc.PATH_PRJ +'prj.json';
+		this.#PATH_APP_JS = this.pc.PATH_WS +'/doc/app.js';
+		this.#PATH_PKG_JSON = this.pc.PATH_WS +'/package.json';
 
-		this.#pnlWVFolder	= new WPFolder(
-			ctx,
-			this.#PATH_WS,
-			this.#PATH_PRJ,
-			diff,
-		);
-		this.#stgSn = new WfbSettingSn(sendRequest2LSP);
+		this.#pnlWVFolder = new WPFolder(this.pc);
+		this.#stgSn = new WfbSettingSn(this.pc, reqPrj2LSP);
 
-		this.#PATH_README4FREEM = this.#PATH_WS +'/build/include/readme.txt';
-		this.#PATH_INS_NSH = this.#PATH_WS +'/build/installer.nsh';
-		this.#PATH_ICON = this.#PATH_WS +'/build/icon.png';
+		this.#PATH_README4FREEM = this.pc.PATH_WS +'/build/include/readme.txt';
+		this.#PATH_INS_NSH = this.pc.PATH_WS +'/build/installer.nsh';
+		this.#PATH_ICON = this.pc.PATH_WS +'/build/icon.png';
 
-//		setEscape();	// 非同期禁止
+		// setEscape();	// 非同期禁止
 
-		this.#setEscape = ()=> sendRequest2LSP('need_go');
+		this.#setEscape = ()=> reqPrj2LSP({cmd: 'need_go'});
 
-		const path_ext = ctx.extensionPath;
+		const path_ext = this.pc.ctx.extensionPath;
 		const path_vue_root = path_ext +'/dist/';
-		this.#localExtensionResRoots = Uri.file(path_vue_root);
-		Promise.allSettled([
-			async ()=> {
+		this.#uriRes = Uri.file(path_vue_root);
+		void Promise.allSettled([
+			()=> Promise.try(()=> {
 				setTitle(cfg.oCfg.book.title);
-				PrjSetting.#hWsFld2token[wsFld.uri.path] = ()=> cfg.oCfg.debuger_token;
-			},
+				PrjSetting.#hWsFld2token[this.pc.wsFld.uri.path] = ()=> cfg.oCfg.debuger_token;
+			}),
 
 			// 設定画面
 			async ()=> {
@@ -140,7 +129,7 @@ export class PrjSetting implements Disposable {
 					path_ext +'/res/img/icon.png', this.#PATH_ICON
 				);
 
-				const fnLaunchJs = this.#PATH_WS +'/.vscode/launch.json';
+				const fnLaunchJs = this.pc.PATH_WS +'/.vscode/launch.json';
 				if (! existsSync(fnLaunchJs)) await copyFile(
 					path_ext +'/res/launch.json', fnLaunchJs
 				);
@@ -158,18 +147,19 @@ export class PrjSetting implements Disposable {
 	]);
 
 	//MARK: デストラクタ
-//	[Symbol.dispose]() {this.#ds.dispose()}
 	dispose() {for (const d of this.#ds) d.dispose()}
 
 
 	getLocalSNVer(): {verSN: string, verTemp: string} {
+		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 		const oPkg = readJsonSync(this.#PATH_PKG_JSON, {encoding: 'utf8'});
-		const fnCngLog = this.#PATH_WS +'/CHANGELOG.md';
-		const lib_name = `@famibee/skynovel${this.is_new_tmp ?'_esm': ''}`;
+		const fnCngLog = this.pc.PATH_WS +'/CHANGELOG.md';
+		const lib_name = `@famibee/skynovel${this.pc.IS_NEW_TMP ?'_esm': ''}`;
 		return {
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
 			verSN	: oPkg.dependencies?.[lib_name]?.slice(1) ?? '',
 			verTemp	: existsSync(fnCngLog)
-				? readFileSync(fnCngLog, {encoding: 'utf8'}).match(/## v(.+)\s/)?.[1] ?? ''
+				? (/## v(.+)\s/.exec(readFileSync(fnCngLog, {encoding: 'utf8'})))?.[1] ?? ''
 				: '',
 		};
 	}
@@ -180,16 +170,17 @@ export class PrjSetting implements Disposable {
 		return PrjSetting.#hWsFld2token[wsFld.uri.path]?.() ?? '';
 	}
 
-	onCreDir({path}: Uri) {
+	async onCreDir({path}: Uri) {
 		this.cfg.oCfg.code[basename(path)] = false;
 		//fs.outputJson(this.fnPrjJs, this.oCfg);
 			// これを有効にすると（Cre & Del）時にファイルが壊れるので省略
-		this.cmd2Vue(<T_E2V_CFG>{cmd: 'update.oCfg', oCfg: this.cfg.oCfg});
+		await this.cmd2Vue(<T_E2V_CFG>{cmd: 'update.oCfg', oCfg: this.cfg.oCfg});
 	}
-	onDelDir({path}: Uri) {
+	async onDelDir({path}: Uri) {
+		// eslint-disable-next-line @typescript-eslint/no-dynamic-delete
 		delete this.cfg.oCfg.code[basename(path)];
-		this.#writePrjJs();
-		this.cmd2Vue(<T_E2V_CFG>{cmd: 'update.oCfg', oCfg: this.cfg.oCfg});
+		void this.#writePrjJs();
+		await this.cmd2Vue(<T_E2V_CFG>{cmd: 'update.oCfg', oCfg: this.cfg.oCfg});
 
 		this.#pnlWVFolder.close();
 	}
@@ -204,162 +195,182 @@ export class PrjSetting implements Disposable {
 
 
 	//MARK: 設定ビューを開く
-	#pnlWV	: WebviewPanel | undefined = undefined;
-	get isOpenPnlWV() {return !!this.#pnlWV}
-	cmd2Vue = (_mes: T_E2V)=> {};
-	#wvuWs	: Uri;
-	get	wvuWs() {return this.#wvuWs}
+	#wp	: WebviewPanel | undefined = undefined;
 	#pathIcon	: string;
 	open() {
 		if (! ActivityBar.aReady[eTreeEnv.NPM]) return;
-
 		const column = window.activeTextEditor?.viewColumn;
-		if (this.#pnlWV) {
-			this.#pnlWV.reveal(column);
-			this.#openSub();
+		if (this.#wp) {
+			this.#wp.reveal(column);
+
+			const wv = this.#wp.webview;
+			wv.html = repWvUri(this.#htmSrc, wv, this.#uriRes);
 			return;
 		}
-
-		const p = this.#pnlWV = window.createWebviewPanel('SKYNovel-setting', '設定', column || ViewColumn.One, {
+		const p = this.#wp = window.createWebviewPanel('SKYNovel-setting', '設定', column ?? ViewColumn.One, {
 			enableScripts		: true,
 		//	retainContextWhenHidden: true,	// 楽だがメモリオーバーヘッド高らしい
 			localResourceRoots	: [
-				this.#localExtensionResRoots,
-				Uri.file(this.#PATH_WS),
+				this.#uriRes,
+				Uri.file(this.pc.PATH_WS),
 			],
 		});
-		p.onDidDispose(()=> this.#pnlWV = undefined, undefined, this.ctx.subscriptions);
-		const wv = this.#pnlWV!.webview;
-		this.#wvuWs = wv.asWebviewUri(Uri.file(this.#PATH_WS));
-		this.#pathIcon = `${this.#wvuWs}/build/icon.png`;
-
-		this.cmd2Vue = (mes: any)=> wv.postMessage(mes);
-		wv.onDidReceiveMessage(m=> this.#hCmd2ProcInput[m.cmd]?.(m), undefined, this.ctx.subscriptions);
-		this.#openSub();
+		p.onDidDispose(()=> {
+			this.#wp = undefined;
+			this.cmd2Vue = _=> Promise.resolve(false);
+		}, undefined, this.pc.ctx.subscriptions);
+		const wv = p.webview;
+		this.#wvuWs = wv.asWebviewUri(Uri.file(this.pc.PATH_WS));
+		this.#pathIcon = `${String(this.#wvuWs)}/build/icon.png`;	// 必ず String() で
+		this.cmd2Vue = async mes=> wv.postMessage(mes);
+		wv.onDidReceiveMessage((m: T_V2E)=> this.#onDidReceiveMessage(m), undefined, this.pc.ctx.subscriptions);
+		wv.html = repWvUri(this.#htmSrc, wv, this.#uriRes);
 	}
-		#openSub() {
-			const wv = this.#pnlWV!.webview;
-			wv.html = this.#htmSrc
-			.replaceAll('${webview.cspSource}', wv.cspSource)
-			.replaceAll(/(href|src)="\.\//g, `$1="${wv.asWebviewUri(this.#localExtensionResRoots)}`);
-		}
+		cmd2Vue: (mes: T_E2V)=> Promise<boolean> = _=> Promise.resolve(false);
+
+		#wvuWs	: Uri;
+		get	wvuWs() {return String(this.#wvuWs)}	// 必ず String() で
 
 	//MARK: パネル入力の対応
-	readonly #hCmd2ProcInput: {[cmd: string]: (m: any)=> Promise<void>}	= {
-		'?'	: async ()=> {
-			this.cmd2Vue(<T_E2V_INIT>{
-				cmd		: '!',
-				oCfg	: this.cfg.oCfg,
-				oWss	: this.#oWss,
-				pathIcon: this.#pathIcon,
-				fld_src	: this.FLD_SRC,
-			});
-			await this.optPic.disp();
-			await this.optSnd.disp();
-			await this.optFont.disp();
-			this.#stgSn.chkMultiMatch();
-		},
+	async #onDidReceiveMessage(m: T_V2E) {switch (m.cmd) {
+		case '?':
+			await Promise.allSettled([
+				(async ()=> {
+					await this.cmd2Vue(<T_E2V_INIT>{
+						cmd		: '!',
+						oCfg	: this.cfg.oCfg,
+						oWss	: this.#oWss,
+						pathIcon: this.#pathIcon,
+						fld_src	: this.pc.FLD_SRC,
+					});
+					this.#stgSn.chkMultiMatch();
+				})(),
+				this.optFont.disp(),
+				this.optPic.disp(),
+				this.optSnd.disp(),
+			]);
+			break;
 
-		'update.oCfg'	: async (e: T_E2V_CFG)=> {
+		case 'update.oCfg':{
+			const {oCfg}: T_V2E_CFG = m;
 			const escOld = this.cfg.oCfg.init.escape;
 			// コピー
 			const oc = this.cfg.oCfg;
-			const save_ns = e.oCfg.save_ns ?? oc.save_ns;
+			const save_ns = oCfg.save_ns ?? oc.save_ns;
 			const c = this.cfg.oCfg = {
 				...oc,
-				book	: {...oc.book, ...e.oCfg.book},
+				book	: {...oc.book, ...oCfg.book},
 				save_ns,
-				window	: {...oc.window, ...e.oCfg.window},
-				log		: {...oc.log, ...e.oCfg.log},
-				init	: {...oc.init, ...e.oCfg.init},
-				debug	: {...oc.debug, ...e.oCfg.debug},
-				code	: {...oc.code, ...e.oCfg.code},
+				window	: {...oc.window, ...oCfg.window},
+				log		: {...oc.log, ...oCfg.log},
+				init	: {...oc.init, ...oCfg.init},
+				debug	: {...oc.debug, ...oCfg.debug},
+				code	: {...oc.code, ...oCfg.code},
 				debuger_token	:
 					this.#setTmp_save_ns.has(save_ns)
 					? ''	// テンプレのままなら空白とする（使用者のみセットさせる）
-					: (e.oCfg.debuger_token ?? oc.debuger_token) || randomUUID(),
+					: (oCfg.debuger_token ?? oc.debuger_token) || randomUUID(),
 			};
 
-			if (c.init.escape !== escOld) this.#setEscape();
+			if (c.init.escape !== escOld) await this.#setEscape();
 			await this.#writePrjJs();
 
 			this.setTitle(c.book.title);
 
 			// package.json
 			const CopyrightYear = new Date().getFullYear();
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 			const p = await readJson(this.#PATH_PKG_JSON, {encoding: 'utf8'});
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.name = c.save_ns;
-			p.appBundleId = p.appId
-				= `com.fc2.blog.famibee.skynovel.${c.save_ns}`;
-			
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+			p.appBundleId = p.appId = `com.fc2.blog.famibee.skynovel.${c.save_ns}`;
+
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.version = c.book.version;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.productName = c.book.title;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.author = c.book.creator;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.appCopyright = `(c)${c.book.creator}`;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.homepage = c.book.pub_url;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.description = c.book.detail;
-			
+
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
 			p.build.appId = p.appId;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 			p.build.productName = c.book.title.normalize('NFD');
 			// p.build.productName = c.book.title;
 				// electron-builder 不具合対策
 				// macOS app crashes when build.productName contains NFC characters · Issue #9264 · electron-userland/electron-builder https://github.com/electron-userland/electron-builder/issues/9264
-			p.build.artifactName = `\${name}-\${version}-\${arch}.\${ext}`;
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+			p.build.artifactName = '${name}-${version}-${arch}.${ext}';
 			await writeFile(this.#PATH_PKG_JSON, JSON.stringify(p, null, '\t'));
 
 			// src/main/main.ts, doc/app.js
-			if (this.is_new_tmp) replaceRegsFile(this.#PATH_WS +'/src/main/main.ts', [
-				[
-					/(companyName\s*:\s*)(['"]).*\2/,
-					`$1"${c.book.publisher}"`
-				],
-				[	// ついでに発表年を
-					/(pkg.appCopyright \+' )\d+/,
-					`$1${CopyrightYear}`
-				],
+			if (this.pc.IS_NEW_TMP) replaceRegsFile(this.pc.PATH_WS +'/src/main/main.ts', [
+			[
+				/(companyName\s*:\s*)(['"]).*\2/,
+				`$1"${c.book.publisher}"`
+			],
+			[	// ついでに発表年を
+				/(pkg.appCopyright \+' )\d+/,
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+				`$1${CopyrightYear}`
+			],
 			], false);
 			else replaceRegsFile(this.#PATH_APP_JS, [
-				[
-					/(companyName\s*:\s*)(['"]).*\2/,
-					`$1"${c.book.publisher}"`
-				],
-				[	// ついでに発表年を
-					/(pkg.appCopyright \+' )\d+/,
-					`$1${CopyrightYear}`
-				],
+			[
+				/(companyName\s*:\s*)(['"]).*\2/,
+				`$1"${c.book.publisher}"`
+			],
+			[	// ついでに発表年を
+				/(pkg.appCopyright \+' )\d+/,
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+				`$1${CopyrightYear}`
+			],
 			], false);
 
 			// build/include/readme.txt
 			replaceRegsFile(this.#PATH_README4FREEM, [
-				[/(【Version】)[^\n]+/g, `$1${c.book.version}`],
-				[/(【タイトル】)[^\n]+/g, `$1${c.book.title}`],
-				[/(【著 作 者】)[^\n]+/g, `$1${c.book.creator}`],
-				[/(【連 絡 先】メール： )[^\n]+/, `$1${c.book.cre_url}`],
-				[	// ついでに発表年を
-					/(Copyright \(C\) )\d+ "([^"]+)"/g,
-					`$1${CopyrightYear} "${c.book.publisher}"`
-				],
-				[/(　　　　　　ＷＥＢ： )[^\n]+/g, `$1${c.book.pub_url}`],
+			[/(【Version】)[^\n]+/g, `$1${c.book.version}`],
+			[/(【タイトル】)[^\n]+/g, `$1${c.book.title}`],
+			[/(【著 作 者】)[^\n]+/g, `$1${c.book.creator}`],
+			[/(【連 絡 先】メール： )[^\n]+/, `$1${c.book.cre_url}`],
+			[	// ついでに発表年を
+				/(Copyright \(C\) )\d+ "([^"]+)"/g,
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+				`$1${CopyrightYear} "${c.book.publisher}"`
+			],
+			// eslint-disable-next-line no-irregular-whitespace
+			[/(　　　　　　ＷＥＢ： )[^\n]+/g, `$1${c.book.pub_url}`],
 			], false);
 
 			// build/installer.nsh
 			replaceRegsFile(this.#PATH_INS_NSH, [
 				[/(!define PUBLISHER ").+"/, `$1${c.book.publisher}"`],
 			], false);
-		},
 
-		'change.range.webp_q_def': m=> this.optPic.chgWebp_q_def(m),
+		}	break;
 
-		'change.range.webp_q'	: m=> this.optPic.chgWebp_q(m),
+		case 'change.range.webp_q_def':
+			await this.optPic.chgWebp_q_def(m);
+			break;
 
-		// views/store/stWSS.ts .cmd2Ex() からのイベント
-		'update.oWss'	: async (e: T_V2E_WSS)=> {
-			for (const [id, val] of Object.entries(e.oWss)) {
-				const old_val = (<any>this.#oWss)[id];
-				if (old_val == val) continue;
+		case 'change.range.webp_q':
+			await this.optPic.chgWebp_q(m);
+			break;
+
+		case 'update.oWss':
+			for (const [id, val] of Object.entries(m.oWss)) {
+				const old_val = <never>this.#oWss[<TK_WSS>id];
+				if (old_val === val) continue;
 
 				await this.#wss.update(id, val);
-				(<any>this.#oWss)[id] = val;
+				this.#oWss[<TK_WSS>id] = <never>val;
 				switch (id) {
 					case 'cnv.font.subset':		break;
 				//	case 'cnv.icon.shape':		continue;
@@ -371,64 +382,54 @@ export class PrjSetting implements Disposable {
 				}
 
 				const o: T_E2V_NOTICE_COMPONENT = {cmd: 'notice.Component', id, mode: 'wait'};
-				this.cmd2Vue(o);
+				await this.cmd2Vue(o);
 
 				const go = await this.#onSettingEvt(id, String(val));
-				if (go) {
-					switch (id) {
-						case 'cnv.mat.pic':
-							await this.optPic.disp();
-							break;
-
-						case 'cnv.mat.snd':
-						case 'cnv.mat.snd.codec':
-							await this.optSnd.disp();
-							break;
-					}
-					o.mode = 'comp';
-				}
+				if (go) o.mode = 'comp';
 				else {
-					await this.#wss.update(id, (<any>this.#oWss[id]) = old_val);
+					await this.#wss.update(id, this.#oWss[<TK_WSS>id] = old_val);
 					o.mode = 'cancel';
 				}
-				this.cmd2Vue(o);
+				await this.cmd2Vue(o);
 				break;
 			}
-		},
+			break;
 
-		'update.aTemp'		: m=> this.#stgSn.update(m),
+		case 'update.aTemp': this.#stgSn.update(m); break;
 
-		'info'	: m=> window.showInformationMessage(m.mes) as Promise<void>,
-		'warn'	: m=> window.showWarningMessage(m.mes) as Promise<void>,
+		case 'info':	window.showInformationMessage(m.mes);	break;
+		case 'warn':	window.showWarningMessage(m.mes);	break;
+		case 'openURL':	openURL(Uri.parse(m.url), this.pc.PATH_WS);	break;
 
-		'openURL': async m=> openURL(Uri.parse(m.url), this.#PATH_WS),
-
-		'copyTxt': async m=> {
-			if (m.id !== 'copy.folder_save_app') return;
+		case 'copyTxt':{
+			if (m.id !== 'copy.folder_save_app') break;
 
 			const {username} = userInfo();
 			switch (process.platform) {
-				case 'win32':
-					await env.clipboard.writeText(`C:\\Users\\${username}\\AppData\\Roaming\\${this.cfg.oCfg.save_ns}\\storage\\`);
-					break;
-				case 'darwin':
-					await env.clipboard.writeText(`/Users/${username}/Library/Application Support/${this.cfg.oCfg.save_ns}/storage/`);
-					break;
-				case 'linux':
-					await env.clipboard.writeText(`~/.config/${this.cfg.oCfg.save_ns}/storage/`);
-					break;
+			case 'win32':
+				await env.clipboard.writeText(`C:\\Users\\${username}\\AppData\\Roaming\\${this.cfg.oCfg.save_ns}\\storage\\`);
+				break;
+			case 'darwin':
+				await env.clipboard.writeText(`/Users/${username}/Library/Application Support/${this.cfg.oCfg.save_ns}/storage/`);
+				break;
+			case 'linux':
+				await env.clipboard.writeText(`~/.config/${this.cfg.oCfg.save_ns}/storage/`);
+				break;
 			}
-			window.showInformationMessage(`クリップボードに【アプリ版（通常実行）セーブデータ保存先パス】をコピーしました`);
-		},
+			window.showInformationMessage('クリップボードに【アプリ版（通常実行）セーブデータ保存先パス】をコピーしました');
+		}	break;
 
-		'selectFile'	: m=> this.optPic.cnvIconShape(m, this.#pathIcon),
-	};
+		case 'selectFile':
+			await this.optPic.cnvIconShape(m, this.#pathIcon);
+			break;
+	}}
+
 
 	//MARK: 設定パネルイベント
 	// 主に設定画面からのアクション。falseを返すとスイッチなどコンポーネントを戻せる
 	async #onSettingEvt(nm: string, val: string): Promise<boolean> {
 //console.log(`fn:Project.ts #cmd nm:${nm} val:${val}`);
-		// 最新は val。this.ctx.workspaceState.get(（など）) は前回値
+		// 最新は val。this.pc.ctx.workspaceState.get(（など）) は前回値
 		switch (nm) {
 		case 'cnv.font.subset':
 			if (await window.showInformationMessage('フォントサイズ最適化（する / しない）を切り替えますか？', {modal: true}, 'はい') !== 'はい') return false;
