@@ -5,36 +5,33 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import {foldProc, fsp2fp} from './CmnLib';
+import {foldProc, repWvUri, vsc2fp} from './CmnLib';
 import {getNonce} from './ActivityBar';
 import {SEARCH_PATH_ARG_EXT} from './ConfigBase';
-import {HDiff} from './HDiff';
+import type {PrjCmn} from './PrjCmn';
 
-import type {ExtensionContext, WebviewPanel} from 'vscode';
+import type {WebviewPanel} from 'vscode';
 import {ViewColumn, Uri, window} from 'vscode';
 import {readFile} from 'fs-extra';
 
 
 export class WPFolder {
-	#wp				: WebviewPanel | undefined	= undefined;
-	#uriOpFolder	: Uri | null				= null;
-	#uriWvPrj		: Uri | null				= null;
+	#wp			: WebviewPanel | undefined	= undefined;
+	#uriOpen	: Uri | null				= null;
+	#uriWvPrj	: Uri | null				= null;
 
-	readonly	#localExtensionResRoots: Uri;
-				#htm	: string;
+	readonly	#uriRes	: Uri;
+				#htmSrc	: string;
 
 
 	//MARK: コンストラクタ
 	constructor(
-		private	readonly	ctx			: ExtensionContext,
-		private	readonly	PATH_WS		: string,
-		private	readonly	PATH_PRJ	: string,
-		private	readonly	diff		: HDiff,
+		private readonly pc		: PrjCmn,
 	) {
-		const path_root = ctx.extensionPath +'/views/';
-		this.#localExtensionResRoots = Uri.file(path_root);
-		readFile(path_root +'folder.htm', {encoding: 'utf8'})
-		.then(t=> this.#htm = t
+		const path_res = this.pc.ctx.extensionPath +'/views/';
+		this.#uriRes = Uri.file(path_res);
+		void readFile(path_res +'folder.htm', {encoding: 'utf8'})
+		.then(t=> this.#htmSrc = t
 			.replace('<meta_autooff ', '<meta ')// ローカルデバッグしたいので
 			.replaceAll('${nonce}', getNonce())
 			.replace('.ts"></script>', '.js"></script>')
@@ -47,57 +44,58 @@ export class WPFolder {
 		// フォルダビュー
 		const column = window.activeTextEditor?.viewColumn;
 		const wp = this.#wp;
-		if (this.#uriOpFolder === uri && wp) {wp.reveal(column); return;}
+		if (this.#uriOpen === uri && wp) {wp.reveal(column); return;}
 
 		if (! wp) {
-			const wp = this.#wp = window.createWebviewPanel('SKYNovel-folder', '', column || ViewColumn.One, {
+			const wp = this.#wp = window.createWebviewPanel('SKYNovel-folder', '', column ?? ViewColumn.One, {
 				enableScripts		: true,
 			//	retainContextWhenHidden: true,// 楽だがメモリオーバーヘッド高らしい
 				localResourceRoots	: [
-					this.#localExtensionResRoots,
-					Uri.file(this.PATH_WS),
+					this.#uriRes,
+					Uri.file(this.pc.PATH_WS),
 				],
 			});
 			const wv = wp.webview;
-			this.ctx.subscriptions.push(
-				wp.onDidDispose(()=> this.#wp = undefined, undefined, this.ctx.subscriptions),	// 閉じられたとき
+			this.pc.ctx.subscriptions.push(
+				wp.onDidDispose(()=> this.#wp = undefined, undefined, this.pc.ctx.subscriptions),	// 閉じられたとき
 
-				wv.onDidReceiveMessage(m=> {switch (m.cmd) {
-					case 'info': window.showInformationMessage(m.text); break;
-					case 'warn': window.showWarningMessage(m.text); break;
+				wv.onDidReceiveMessage(({cmd, text}: {cmd: string, text: string})=> {switch (cmd) {
+					case 'info': window.showInformationMessage(text); break;
+					case 'warn': window.showWarningMessage(text); break;
 				}}, false),
 			);
 
-			wv.html = this.#htm
-				.replaceAll('${webview.cspSource}', wv.cspSource)
-				.replaceAll(/(href|src)="\.\//g, `$1="${wv.asWebviewUri(this.#localExtensionResRoots)}/`)
-				.replace(/<!--SOL-->.+?<!--EOL-->/s, '');
-					// https://regex101.com/r/8RaTsD/1
+			wv.html = repWvUri(this.#htmSrc, wv, this.#uriRes)
+			.replace(/<!--SOL-->.+?<!--EOL-->/s, '');
+				// https://regex101.com/r/8RaTsD/1
 		}
-		this.#uriOpFolder = uri;
+		this.#uriOpen = uri;
 
 		this.#update(uri);
 	}
 	//MARK: ビュー更新
 	#update(uri: Uri) {
 		const vfp = uri.path;
-		const fp = fsp2fp(vfp);
-		const pp = this.diff.fp2pp(fp);
+		const fp = vsc2fp(vfp);
+		const pp = this.pc.diff.fp2pp(fp);
+		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 		this.#wp!.title = pp +' フォルダ';
 
 		let htm = '';
+		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 		const wv = this.#wp!.webview;
-		this.#uriWvPrj = wv.asWebviewUri(Uri.file(this.PATH_PRJ));
+		this.#uriWvPrj = wv.asWebviewUri(Uri.file(this.pc.PATH_PRJ));
+		const pathWp = String(this.#uriWvPrj);	// 必ず String() で
 		foldProc(fp, (vfp2, nm)=> {
-			const fp2 = fsp2fp(Uri.file(vfp2).path);
-			const pp2 = this.diff.fp2pp(fp2);
+			const fp2 = vsc2fp(Uri.file(vfp2).path);
+			const pp2 = this.pc.diff.fp2pp(fp2);
 			if (this.#REG_MOV.test(fp2)) {	// GrpよりMovを先に
 				htm +=
 `<div class="col pe-0">
 	<div class="card text-bg-secondary">
-		<video controls controlsList="nodownload" playsinline preload="metadata" class="w-100" src="${this.#uriWvPrj}${pp2}" class="card-img-top"></video>
+		<video controls controlsList="nodownload" playsinline preload="metadata" class="w-100" src="${pathWp}${pp2}" class="card-img-top"></video>
 		<div class="card-body">
-			<a href="${this.#uriWvPrj}${pp2}" title="エディタにドラッグできます" class="btn btn-primary" title="エディタにドラッグできます">${nm}</a>
+			<a href="${pathWp}${pp2}" title="エディタにドラッグできます" class="btn btn-primary" title="エディタにドラッグできます">${nm}</a>
 		</div>
 	</div>
 </div>`;
@@ -107,7 +105,7 @@ export class WPFolder {
 			if (this.#REG_GRP.test(pp2)) {
 				htm +=
 `<div class="col pe-0">
-	<img loading="lazy" src="${this.#uriWvPrj}${pp2}" title="${nm}"/>
+	<img loading="lazy" src="${pathWp}${pp2}" title="${nm}"/>
 </div>`;
 				return;
 			}
@@ -116,29 +114,29 @@ export class WPFolder {
 				htm +=
 `<div class="col pe-0">
 	<div class="card text-bg-secondary">
-		<audio controls src="${this.#uriWvPrj}${pp2}" class="card-img-top"></audio>
+		<audio controls src="${pathWp}${pp2}" class="card-img-top"></audio>
 		<div class="card-body">
-			<a href="${this.#uriWvPrj}${pp2}" title="エディタにドラッグできます" class="btn btn-primary" title="エディタにドラッグできます">${nm}</a>
+			<a href="${pathWp}${pp2}" title="エディタにドラッグできます" class="btn btn-primary" title="エディタにドラッグできます">${nm}</a>
 		</div>
 	</div>
 </div>`;
 				return;
 			}
 
-		}, ()=> {});
+		}, ()=> { /* empty */ });
 		wv.postMessage({cmd: 'refresh', o: {htm}});
 	}
 		readonly #REG_MOV = /\.(mp4|webm)$/;
-		// (new RegExp("~")) の場合は、バックスラッシュは２つ必要
+		// (new RegExp('\')) の場合は、バックスラッシュは２つ必要
 		readonly #REG_GRP = new RegExp(`\\.${SEARCH_PATH_ARG_EXT.SP_GSM}$`);
 		readonly #REG_SND = new RegExp(`\\.${SEARCH_PATH_ARG_EXT.SOUND}$`);
 
 	//MARK: ビュー遅延更新
 	updateDelay({path}: Uri) {
-		const uri = this.#uriOpFolder;
+		const uri = this.#uriOpen;
 		if (! this.#wp || ! uri) return;
-		const fp = fsp2fp(path);			// /c:/
-		const fpOF = fsp2fp(uri.path);	// /C:/
+		const fp = vsc2fp(path);			// /c:/
+		const fpOF = vsc2fp(uri.path);	// /C:/
 		if (! fp.startsWith(fpOF)) return;
 
 		if (this.#tiDelay) clearTimeout(this.#tiDelay);	// 遅延
@@ -159,7 +157,7 @@ export class WPFolder {
 	close() {
 		this.#wp?.dispose();
 		this.#wp = undefined;
-		this.#uriOpFolder = null;
+		this.#uriOpen = null;
 	}
 
 }
