@@ -5,51 +5,59 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-const [, , src, shape, path, is_new_tmp] = process.argv;
+import type {T_BJ_cut_round} from '../types';
 
 import sharp from 'sharp';
 sharp.cache(false);
-
-import {styleText} from 'node:util';
-import {existsSync} from 'node:fs';
-import {readFile, stat, writeFile} from 'node:fs/promises';
-import {fileURLToPath} from 'node:url';
-import {copy, mkdirs, writeJsonSync} from 'fs-extra/esm';
 import {BICUBIC2, BILINEAR, createICNS, createICO} from 'png2icons';
 
+import {fileURLToPath} from 'node:url';
+import {styleText} from 'node:util';
+import {existsSync} from 'node:fs';
+import {readFile, writeFile} from 'node:fs/promises';
+import {copy, mkdirs, readJson, writeJsonSync} from 'fs-extra/esm';
+
 const __filename = fileURLToPath(import.meta.url);
+const fnBJ = __filename +'on';
+const PATH_WORKSPACE = process.cwd() +'/';
 
-sharp(src).metadata().then(async meta=> {
-	const oLog: any = {...meta, err: '', exif: '', icc: '', iptc: '', xmp: ''};
-	const log_exit = (exit_code = -1)=> {
-		writeJsonSync(__filename +'on', oLog, {encoding: 'utf8'});
-		if (exit_code > -1) process.exit(exit_code);
-	}
-	log_exit();
+function exit(mes: string, exit_code = 0) {
+	if (exit_code === 0)
+	   console.log(styleText(['bgGreen', 'black'], '  [OK] %o'), mes);
+	else console.log(styleText(['bgRed', 'white'], '  [ERR] %o'), mes);
 
-	if (! meta || ! meta.width || ! meta.height) {
-		console.log(styleText(['bgRed', 'white'], `  画像のサイズが不明です`));
-		log_exit(20);
-		return;
+	process.exit(exit_code);
+}
+
+
+readJson(fnBJ, {encoding: 'utf8'})
+.then(async (oBJ: T_BJ_cut_round)=> {try {
+	function log_exit(mes: string, exit_code = 0) {
+		writeJsonSync(fnBJ, oBJ, {encoding: 'utf8'});
+		exit(mes, exit_code);
 	}
+
+	const {order} = oBJ;
+	const sh = sharp((order.is_src_pp ?PATH_WORKSPACE :'') + order.wp_src);
+	const {width, height} = await sh.metadata();
+	if (! width || ! height) throw '画像のサイズが不明です';
 
 	const wh = 1024;
-	if (meta.width < wh || meta.height < wh) {
-		oLog.err = `元画像のサイズは ${wh} x ${wh} 以上にして下さい。（width:${meta.width} height:${meta.height}）`;
-		log_exit(10);
+	if (width < wh || height < wh) {
+		// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+		throw `元画像のサイズは ${wh} x ${wh} 以上にして下さい。（width:${width} height:${height}）`;
 	}
-
-	const pathWs = process.cwd() +'/';
-	const s = sharp(src).png().resize({
+	const s = sh.png().resize({
 		width	: wh,
 		height	: wh,
 		fit		: 'cover',
 		background	: {r: 0, g: 0, b: 0, alpha: 0},
 	});
-
-	switch (parseInt(shape!)) {
+	// eslint-disable-next-line @typescript-eslint/no-unnecessary-type-conversion
+	switch (Number(order.shape)) {	// JSON がうまく number にできないので
 		case 1:		// 丸
 			s.composite([{
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
 				input	: Buffer.from(`<svg><circle cx="${wh /2}" cy="${wh /2}" r="${wh /2}"/></svg>`),
 				blend	: 'dest-in',
 			}]);
@@ -57,6 +65,7 @@ sharp(src).metadata().then(async meta=> {
 
 		case 2:		// 角丸
 			s.composite([{
+				// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
 				input	: Buffer.from(`<svg><rect width="${wh}" height="${wh}" rx="${wh/4.5}" ry="${wh/4.5}"/></svg>`),
 				blend	: 'dest-in',
 			}]);
@@ -65,45 +74,31 @@ sharp(src).metadata().then(async meta=> {
 	//	default:
 	}
 
-	try {
-		await s.toFile(pathWs + path);
+	await s.toFile(PATH_WORKSPACE + order.wp_dest);
 
-		const fnIcon = pathWs +'build/icon.png';
-		if (! existsSync(fnIcon)) return;
+	const fnIcon = PATH_WORKSPACE +'build/icon.png';
+	if (! existsSync(fnIcon)) throw '生成に失敗しました';
 
-		// サムネイル更新
-		const mtPng = (await stat(fnIcon)).mtimeMs;
-		const bIconPng = await readFile(fnIcon);
-		await mkdirs(pathWs +'build/icon/');
+	// サムネイル更新
+	const bIconPng = await readFile(fnIcon);
+	await mkdirs(PATH_WORKSPACE +'build/icon/');
 
-		await Promise.allSettled([
-			async ()=> {
-				const fn = pathWs +'build/icon/icon.icns';
-				const mt = existsSync(fn) ?(await stat(fn)).mtimeMs :0;
-				if (mtPng > mt) {
-					const b = createICNS(bIconPng, BILINEAR, 0);
-					if (b) await writeFile(fn, b);
-				}
-			},
-			async ()=> {
-				const fn = pathWs +'build/icon/icon.ico';
-				const mt = existsSync(fn) ?(await stat(fn)).mtimeMs :0;
-				if (mtPng > mt) {
-					const b = createICO(bIconPng, BICUBIC2, 0, false, true);
-					if (b) await writeFile(fn, b);
-				}
-			},
-			// 「このアプリについて」用
-			()=> copy(fnIcon, pathWs +is_new_tmp ?'doc/icon.png' :'doc/app/icon.png'),
-		].map(v=> v()));
-	} catch (e) {
-		console.log(styleText(['bgRed', 'white'], `  [ERR] %o`), e);
-		oLog.err = (<Error>e)?.message ?? String(e);
-		log_exit(20);
-		return;
-	}
+	await Promise.allSettled([
+		(async ()=> {
+			const fn = PATH_WORKSPACE +'build/icon/icon.icns';
+			const b = createICNS(bIconPng, BILINEAR, 0);
+			if (b) await writeFile(fn, b);
+		})(),
+		(async ()=> {
+			const fn = PATH_WORKSPACE +'build/icon/icon.ico';
+			const b = createICO(bIconPng, BICUBIC2, 0, false, true);
+			if (b) await writeFile(fn, b);
+		})(),
+		// 「このアプリについて」用
+		copy(fnIcon, PATH_WORKSPACE +`doc/${ order.is_new_tmp ?'' :'app/' }icon.png`),
+	]);
 
-	console.log(styleText(['bgGreen', 'black'], `fn:cut_round.ts ok.`));
-	process.exit(0);
+	log_exit('終了しました');
 
-});
+} catch (e) {exit(oBJ.err = String(e), 20)}})
+.catch((e: unknown)=> {exit(String(e), 10)});
