@@ -8,35 +8,37 @@
 import {defineStore} from 'pinia';
 import {useCfg} from '../store/stCfg';
 import {useWss} from '../store/stWSS';
-import type {T_E2V_INIT} from '../types';
+import type {T_E2V_INIT, T_E2V, T_Ex2Vue_cmd, T_V2E_Q} from '../../src/types';
+// eslint-disable-next-line no-undef
 const vscode = ('acquireVsCodeApi' in window) ?acquireVsCodeApi() :undefined;
 export const isVSCode = vscode !== undefined;
 
-export const cmd2Ex: (o: any)=> void = vscode
+export const cmd2Ex: (o: unknown)=> void = vscode
 	? o=> vscode.postMessage(o)
-	: o=> console.log(`cmd2Ex:%o`, o);
+	: o=> console.log('cmd2Ex:%o', o);
 export const info = (mes: string)=> cmd2Ex({cmd: 'info', mes});
 export const warn = (mes: string)=> cmd2Ex({cmd: 'warn', mes});
 export const openURL = (url: string)=> cmd2Ex({cmd: 'openURL', url});
 export const copyTxt = (id: string)=> cmd2Ex({cmd: 'copyTxt', id});
 
 
-const aHook: {nm: string; fnc: (m: any)=> void;}[] = [];
-export const on = (nm: string, fnc: (data: any)=> void)=> aHook.push({nm, fnc});
+const hHook: {[nm: string]: ((d: T_E2V)=> void)[]} = {};
+export function on(nm: T_Ex2Vue_cmd, fnc: (d: any)=> void) {
+	(hHook[nm] ??= []).push(fnc);
+}
 window.addEventListener('message', e=> {
-	if (! e.isTrusted) {warn('Setting.vue isTrusted=false'); return;}
-	go(e.data.cmd, e.data);
+	if (! e.isTrusted) {warn('Setting.vue isTrusted=false'); return}
+	go(<T_E2V>e.data);
 });
-function go(nm: string, data: any) {for (const v of aHook) {
-	if (v.nm === nm) v.fnc(data);
-}}
-cmd2Ex({cmd: '?'});	// 拡張機能メインへ準備完了通知
+function go(d: T_E2V) { for (const f of hHook[d.cmd] ?? []) f(d); }
+cmd2Ex(<T_V2E_Q>{cmd: '?'});	// 拡張機能メインへ準備完了通知
 
 
 type T_STATE = {
 	active_tab	: string;
 };
-const oVSCode: T_STATE = vscode?.getState() as T_STATE ?? {	// 永続性回復
+// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+const oVSCode: T_STATE = <T_STATE>vscode?.getState() ?? {	// 永続性回復
 	active_tab	: 'basic',
 };
 
@@ -56,13 +58,14 @@ export const useVSCode = ()=> {
 		st.$subscribe(()=> vscode?.setState(oVSCode));
 		const stCfg = useCfg();
 		// 拡張機能メインから初期値取得	// 必ず st.$subscribe()以後に
-		on('!', ({oCfg, oWss}: T_E2V_INIT)=> {
+		on('!', d=> {
+			const {oCfg, oWss} = <T_E2V_INIT>d;
 			stCfg.init(oCfg);
 
 			const stWss = useWss();
 			stWss.init(oWss);
 
-			go('init', {});
+			go({cmd: 'init.Vue'});
 		});
 	}
 
@@ -70,7 +73,8 @@ export const useVSCode = ()=> {
 };
 
 export const getLeftRangeBadge = (value=0, max=0, min=0)=> {
-	const val = Number( (value - min) *100 /(max - min) );
+	const val = (value - min) *100 /(max - min);
 	const pos = 10 -(val *0.2);
+	// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
 	return `calc(${val}% + (${pos}px))`;
 };
