@@ -56,7 +56,7 @@ import {storeToRefs} from 'pinia';
 import {useField} from 'vee-validate';
 import * as yup from 'yup';
 import {openURL, on} from '../store/stVSCode';
-import type {T_CFG} from '../types';
+import type {T_CFG} from '../../src/types';
 
 const stCfg = useCfg();
 const {oCfg} = storeToRefs(stCfg);
@@ -128,7 +128,7 @@ const {value: v_detail, errorMessage: em_detail, meta: mv_detail} = useField<str
 	{initialValue: oCfg.value.book!.detail},	// ブラウザテスト用、VSCodeで上書き
 );
 
-on('init', ()=> {	// useField()の後に初期値を更新したいので
+on('init.Vue', ()=> {	// useField()の後に初期値を更新したいので
 	const o: T_CFG = oCfg.value;
 	v_save_ns.value = o.save_ns!;
 	v_title.value = o.book!.title!;
